+++ ase/test/dimensionality.py
@@ -1,6 +1,6 @@
-from ase.geometry import analyze_dimensionality
 import ase.build
 from ase.lattice.cubic import FaceCenteredCubic
+from ase.geometry.dimensionality import analyze_dimensionality
 
 
 # 2D test
@@ -10,21 +10,21 @@ atoms.set_pbc((1, 1, 1))
 atoms *= 2
 
 intervals = analyze_dimensionality(atoms, method='RDA')
-(score, a, b, hr, h, components, cdim) = intervals[0]
-assert hr == '2D'
+m = intervals[0]
+assert m.dimtype == '2D'
 
 intervals = analyze_dimensionality(atoms, method='TSA')
-(score, a, b, hr, h, components, cdim) = intervals[0]
-assert hr == '2D'
+m = intervals[0]
+assert m.dimtype == '2D'
 
 
 # 3D test
 atoms = FaceCenteredCubic(size=(2, 2, 2), symbol='Cu', pbc=(1, 1, 1))
 
 intervals = analyze_dimensionality(atoms, method='RDA')
-(score, a, b, hr, h, components, cdim) = intervals[0]
-assert hr == '3D'
+m = intervals[0]
+assert m.dimtype == '3D'
 
 intervals = analyze_dimensionality(atoms, method='TSA')
-(score, a, b, hr, h, components, cdim) = intervals[0]
-assert hr == '3D'
+m = intervals[0]
+assert m.dimtype == '3D'