+++ doc/releasenotes.rst
@@ -17,7 +17,7 @@ Git master branch
 * The genetic algorithm module :mod:`ase.ga` now has operators for crystal
   structure prediction. See :ref:`ga_bulk_tutorial`.
 
-* New :func:`ase.geometry.analyze_dimensionality` function.  See:
+* New :func:`ase.geometry.dimensionality.analyze_dimensionality` function.  See:
   :ref:`dimtutorial`.
 
 * New :func:`ase.utils.deltacodesdft.delta` function:  Calculates the
@@ -26,6 +26,7 @@ Git master branch
 
 * Calculation of stress tensor implemented for
   :class:`~ase.calculators.emt.EMT` potential
+
 
 Version 3.17.0
 ==============